+++ babel.config.js
@@ -7,10 +7,12 @@ module.exports = {
       {
         root: ['.'],
         alias: {
+          'src': './src',
           'app-util': './src/util',
           'app-store': './src/store',
           'app-config': './src/config',
           'app-packages': './src/packages',
+          'app-helper': './src/helper'
         },
       },
     ],