+++ src/components/Social/index.js
@@ -1,0 +1,3 @@
+import ListFeeds from './ListFeeds';
+
+export {ListFeeds};